Refactor: Moved dto from general to general/dto
