+++ api/src/main/java/ua/edu/chdtu/deanoffice/api/course/dto/CourseDTO.java
@@ -3,7 +3,7 @@ package ua.edu.chdtu.deanoffice.api.course.dto;
 import com.fasterxml.jackson.annotation.JsonView;
 import lombok.Getter;
 import lombok.Setter;
-import ua.edu.chdtu.deanoffice.api.general.NamedDTO;
+import ua.edu.chdtu.deanoffice.api.general.dto.NamedDTO;
 
 @Getter
 @Setter
