+++ api/src/main/java/ua/edu/chdtu/deanoffice/api/general/dto/GeneralView.java
@@ -1,4 +1,4 @@
-package ua.edu.chdtu.deanoffice.api.general;
+package ua.edu.chdtu.deanoffice.api.general.dto;
 
 public class GeneralView {
     public interface PersonFullName {}
