+++ api/src/main/java/ua/edu/chdtu/deanoffice/api/general/dto/KnowledgeControlDTO.java
@@ -1,4 +1,4 @@
-package ua.edu.chdtu.deanoffice.api.general;
+package ua.edu.chdtu.deanoffice.api.general.dto;
 
 import lombok.Getter;
 import lombok.Setter;
