+++ api/src/main/java/ua/edu/chdtu/deanoffice/api/group/dto/StudentDegreeFullNameDTO.java
@@ -3,7 +3,7 @@ package ua.edu.chdtu.deanoffice.api.group.dto;
 import com.fasterxml.jackson.annotation.JsonView;
 import lombok.Getter;
 import lombok.Setter;
-import ua.edu.chdtu.deanoffice.api.general.PersonFullNameDTO;
+import ua.edu.chdtu.deanoffice.api.general.dto.PersonFullNameDTO;
 
 @Getter
 @Setter
