+++ api/src/main/java/ua/edu/chdtu/deanoffice/api/group/dto/StudentGroupView.java
@@ -1,6 +1,6 @@
 package ua.edu.chdtu.deanoffice.api.group.dto;
 
-import ua.edu.chdtu.deanoffice.api.general.GeneralView;
+import ua.edu.chdtu.deanoffice.api.general.dto.GeneralView;
 
 public class StudentGroupView {
     public interface Basic {}
