+++ api/src/main/java/ua/edu/chdtu/deanoffice/api/student/dto/StudentDegreeDTO.java
@@ -3,7 +3,7 @@ package ua.edu.chdtu.deanoffice.api.student.dto;
 import com.fasterxml.jackson.annotation.JsonView;
 import lombok.Getter;
 import lombok.Setter;
-import ua.edu.chdtu.deanoffice.api.general.NamedDTO;
+import ua.edu.chdtu.deanoffice.api.general.dto.NamedDTO;
 import ua.edu.chdtu.deanoffice.entity.EducationDocument;
 import ua.edu.chdtu.deanoffice.entity.Payment;
