+++ api/src/main/java/ua/edu/chdtu/deanoffice/api/student/dto/StudentExpelDTO.java
@@ -3,7 +3,7 @@ package ua.edu.chdtu.deanoffice.api.student.dto;
 import com.fasterxml.jackson.annotation.JsonView;
 import lombok.Getter;
 import lombok.Setter;
-import ua.edu.chdtu.deanoffice.api.general.OrderReasonDTO;
+import ua.edu.chdtu.deanoffice.api.general.dto.OrderReasonDTO;
 
 import java.util.Date;
